fix: reorganize tests that rely on network and ignore outside of go test
