+++ cmd/crates/soroban-test/tests/it/integration.rs
@@ -1,0 +1,4 @@
+mod custom_types;
+mod dotenv;
+mod hello_world;
+mod util;
