+++ cmd/crates/soroban-test/tests/it/main.rs
@@ -1,10 +1,7 @@
 mod arg_parsing;
 mod config;
-
-mod custom_types;
-mod dotenv;
-mod hello_world;
-mod integration_and_sandbox;
+#[cfg(ignore)]
+mod integration;
 mod lab;
 mod plugin;
 mod util;
