+++ cmd/soroban-cli/src/commands/contract/restore.rs
@@ -176,7 +176,6 @@ impl Cmd {
         }
         parse_operations(&operations).ok_or(Error::MissingOperationResult)
     }
-
 }
 
 fn parse_operations(ops: &[OperationMeta]) -> Option<u32> {
